requirements.txt
fastapi[all]==0.115.6
alembic==1.14.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pytest==8.3.4
pydantic-settings==2.7.1
asyncpg==0.30.0
geoalchemy2==0.17.0
python-multipart==0.0.20
shapely==2.0.6